requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
Werkzeug==3.0.6
gunicorn==23.0.0
